+++ src/main/java/com/ably/kafka/connect/config/ChannelConfig.java
@@ -1,8 +1,8 @@
-package com.ably.kafka.connect;
+package com.ably.kafka.connect.config;
 
 import io.ably.lib.types.ChannelOptions;
 
-interface ChannelConfig {
+public interface ChannelConfig {
     String getName();
 
     ChannelOptions getOptions() throws ChannelSinkConnectorConfig.ConfigException;
