+++ src/main/java/com/ably/kafka/connect/mapping/DefaultChannelSinkMapping.java
@@ -1,5 +1,8 @@
-package com.ably.kafka.connect;
+package com.ably.kafka.connect.mapping;
 
+import com.ably.kafka.connect.config.ChannelConfig;
+import com.ably.kafka.connect.config.ChannelSinkConnectorConfig;
+import com.ably.kafka.connect.config.ConfigValueEvaluator;
 import io.ably.lib.realtime.AblyRealtime;
 import io.ably.lib.realtime.Channel;
 import io.ably.lib.types.AblyException;
