+++ src/test/java/com/ably/kafka/connect/DefaultChannelSinkMappingTest.java
@@ -1,5 +1,9 @@
 package com.ably.kafka.connect;
 
+import com.ably.kafka.connect.config.ChannelSinkConnectorConfig;
+import com.ably.kafka.connect.config.ConfigValueEvaluator;
+import com.ably.kafka.connect.config.DefaultChannelConfig;
+import com.ably.kafka.connect.mapping.DefaultChannelSinkMapping;
 import io.ably.lib.realtime.AblyRealtime;
 import io.ably.lib.realtime.Channel;
 import io.ably.lib.types.AblyException;
